@@ -5,9 +5,8 @@
   "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "npm run lint && npm run build-ts",
-    "lint": "./node_modules/.bin/tslint --config ./tslint.json './src/**/*.ts'",
-    "build-ts": "./node_modules/.bin/webpack --progress --colors --config ./webpack.config.js"
+    "build": "./node_modules/.bin/webpack --progress --colors --config ./webpack.config.js",
+    "lint": "./node_modules/.bin/tslint --config ./tslint.json './src/**/*.ts'"
   },
   "repository": {
     "type": "git",
@@ -32,11 +31,12 @@
   },
   "homepage": "https://github.com/friedemannsommer/typescript-postmessagehandler#readme",
   "devDependencies": {
-    "json3": "^3.3.2",
-    "ts-loader": "^1.3.0",
-    "tslint": "^4.0.2",
-    "typescript": "^2.0.10",
-    "webpack": "^1.14.0"
+    "awesome-typescript-loader": "5.2.1",
+    "tslint": "5.18.0",
+    "tslint-loader": "3.5.4",
+    "typescript": "3.5.3",
+    "webpack": "4.39.2",
+    "webpack-cli": "3.3.6"
   },
   "dependencies": {}
 }
