@@ -1,16 +1,11 @@
-/// <reference path="./typings/PostMessageEvent.d.ts" />
-/// <reference path="./typings/JSON3.d.ts" />
-
 import addEvent from './lib/add-event'
-import bind from './lib/bind'
+import bind, { BindFn } from './lib/bind'
 import getType from './lib/get-type'
 import isFunction from './lib/is-function'
 import removeEvent from './lib/remove-event'
-import JSON3 = require('json3')
+import { IPostMessageEvent } from './typings/PostMessageEvent'
 
 class PostMessageHandler {
-    private static JSON = JSON3.noConflict()
-
     private static registerListener(func: EventListener): void {
         addEvent(window, 'message', func, false)
     }
@@ -19,29 +14,28 @@ class PostMessageHandler {
         removeEvent(window, 'message', func)
     }
 
-    private static parse(data: string): any[] {
-        return PostMessageHandler.JSON.parse(data)
-    }
-
-    private static serialize(data: any[]): string {
-        return PostMessageHandler.JSON.stringify(data)
-    }
-
+    private readonly messageListener: BindFn[] = []
+    private readonly secret: string
+    private readonly target: MessageEventSource
+    private readonly targetOrigin: string | undefined
+    private readonly eventCallback: EventListener
+    private readonly isWindow: boolean
     private listenerRegistered: boolean = false
-    private messageListener: Function[] = []
-    private secret: string
-    private target: Window
-    private targetOrigin: string
-    private eventCallback: EventListener
 
-    public constructor(secret: string, target: Window, targetOrigin: string) {
+    public constructor(secret: string, target: MessageEventSource, targetOrigin?: string) {
+        this.isWindow = (target as Window).self === target && (target as Window).window === target && typeof window.document === 'object'
+
+        if (this.isWindow && getType(targetOrigin) !== 'string') {
+            throw new Error('`target` is `window` like which requires an `targetOrigin`, but no `targetOrigin` is set')
+        }
+
+        this.targetOrigin = targetOrigin
         this.secret = secret
         this.target = target
-        this.targetOrigin = targetOrigin
-        this.eventCallback = <EventListener>bind(this.handleMessage, this)
+        this.eventCallback = bind(this.handleMessage, this) as EventListener
     }
 
-    public subscribe(func: Function) {
+    public subscribe(func: BindFn) {
         if (!this.listenerRegistered) {
             this.listenerRegistered = true
             PostMessageHandler.registerListener(this.eventCallback)
@@ -52,8 +46,8 @@ class PostMessageHandler {
         }
     }
 
-    public unsubscribe(func: Function) {
-        let index: number = this.messageListener.indexOf(func)
+    public unsubscribe(func: BindFn) {
+        const index: number = this.messageListener.indexOf(func)
 
         if (index > -1) {
             this.messageListener.splice(index, 1)
@@ -65,14 +59,16 @@ class PostMessageHandler {
         }
     }
 
-    public send(...data: any[]): boolean {
-        let userData: string = PostMessageHandler.serialize(data)
-        userData = this.secret + userData
-
+    public send(...data: unknown[]): boolean {
         if (isFunction(this.target.postMessage)) {
             try {
-                this.target.postMessage(userData, this.targetOrigin)
-                return true
+                if (this.isWindow) {
+                    (this.target as Window).postMessage(this.secret + JSON.stringify(data), this.targetOrigin as string)
+                    return true
+                } else {
+                    (this.target as MessagePort).postMessage(this.secret + JSON.stringify(data))
+                    return true
+                }
             } catch (e) {
                 // silence is golden
             }
@@ -85,7 +81,7 @@ class PostMessageHandler {
         return eventOrigin === this.targetOrigin && eventSource === this.target
     }
 
-    private handleMessage(event: PostMessageEvent): void {
+    private handleMessage(event: IPostMessageEvent): void {
         const dataKey: string = (event.message) ? 'message' : 'data'
         const secret: string = (getType(event[dataKey]) === 'string') ? event[dataKey].slice(0, this.secret.length) : ''
         const data: string = (getType(event[dataKey]) === 'string') ? event[dataKey].slice(this.secret.length) : ''
@@ -94,16 +90,16 @@ class PostMessageHandler {
         if (this.checkEventOrigin(event.origin, event.source)) {
             // check if secret match
             if (secret === this.secret) {
-                this.callListener(PostMessageHandler.parse(data))
+                this.callListener(JSON.parse(data))
             }
         }
     }
 
-    private callListener(data: any[]): void {
-        const len: number = this.messageListener.length
+    private callListener(data: unknown[]): void {
+        const length: number = this.messageListener.length
         let index: number = -1
 
-        while (++index < len) {
+        while (++index < length) {
             this.messageListener[index].apply(undefined, data)
         }
     }
