@@ -1,24 +1,27 @@
-import concat from './array-concat'
-import getType from './get-type'
+import isFunction from './is-function'
 
-export default function (fn: Function, context: Object, ...givenArguments: any[]): Function {
-    if (getType(fn) === 'function') {
-        if (getType(Function.prototype.bind) === 'function') {
-            return Function.prototype.bind.apply(fn, concat([context], givenArguments))
+export type BindFn<T extends unknown[] = unknown[]> = (...args: T) => void
+
+export default function bindFn<T extends unknown[] = unknown[]>(
+    fn: BindFn<T>,
+    context: object | undefined,
+    ...givenArguments: unknown[]
+): BindFn<T> {
+    if (isFunction(fn)) {
+        if (isFunction(Function.prototype.bind)) {
+            return Function.prototype.bind.apply(fn, [context].concat(givenArguments) as [object | undefined, ...unknown[]])
         }
 
         return givenArguments.length > 0
-            ?
-            (...localArguments: any[]) => {
+            ? (...localArguments: unknown[]) => {
                 return localArguments.length > 0
-                    ? fn.apply(context, concat(givenArguments, localArguments))
-                    : fn.apply(context, givenArguments)
+                    ? fn.apply(context, givenArguments.concat(localArguments) as T)
+                    : fn.apply(context, givenArguments as T)
             }
-            :
-            (...localArguments: any[]) => {
+            : (...localArguments: unknown[]) => {
                 return localArguments.length > 0
-                    ? fn.apply(context, localArguments)
-                    : fn.apply(context)
+                    ? fn.apply(context, localArguments as T)
+                    : (fn as unknown as BindFn<[]>).call(context)
             }
     }
 
