@@ -1,3 +1,5 @@
-export default function (object: any): string {
-    return Object.prototype.toString.call(object).replace(/^\[object\s+|\]$/ig, '').toLowerCase()
+export default function getType(object: unknown): string {
+    const typeStr = Object.prototype.toString.call(object).substring(8).toLowerCase()
+
+    return typeStr.substring(0, typeStr.length - 1)
 }
