@@ -1,5 +1,5 @@
 import getType from './get-type'
 
-export default function (value: any): boolean {
+export default function isFunction(value: any): boolean {
     return getType(value) === 'function' || typeof value === 'function'
 }
