@@ -1,7 +1,13 @@
-/// <reference path="../typings/EventTarget.d.ts"/>
 import supports from './supports'
 
-export default function (element: EventTarget, type: string, listener: EventListener): void {
+declare global {
+    // tslint:disable-next-line: interface-name
+    interface EventTarget {
+        detachEvent(event: string, listener: EventListener): boolean
+    }
+}
+
+export default function removeEvent(element: EventTarget, type: string, listener: EventListener): void {
     if (supports(element, 'removeEventListener')) {
         element.removeEventListener(type, listener)
     } else if (supports(element, 'detachEvent')) {
